+++ resources/views/components/auth-box.blade.php
@@ -1,5 +1,5 @@
 {{-- auth template layout --}}
-<div class="py-5-3-2 px">
+<div class="py-5-3-2">
     <div class="bx maxw600 mx-auto">
         {{ $slot }}
     </div>