+++ src/routes.php
@@ -1,183 +1,109 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-use Laravel\Fortify\Features;
-use Laravel\Fortify\Http\Controllers\AuthenticatedSessionController;
-use Laravel\Fortify\Http\Controllers\ConfirmablePasswordController;
-use Laravel\Fortify\Http\Controllers\ConfirmedPasswordStatusController;
-use Laravel\Fortify\Http\Controllers\EmailVerificationNotificationController;
-use Laravel\Fortify\Http\Controllers\EmailVerificationPromptController;
-use Laravel\Fortify\Http\Controllers\NewPasswordController;
-use Laravel\Fortify\Http\Controllers\PasswordController;
-use Laravel\Fortify\Http\Controllers\PasswordResetLinkController;
-use Laravel\Fortify\Http\Controllers\ProfileInformationController;
-use Laravel\Fortify\Http\Controllers\RecoveryCodeController;
-use Laravel\Fortify\Http\Controllers\RegisteredUserController;
-use Laravel\Fortify\Http\Controllers\TwoFactorAuthenticatedSessionController;
-use Laravel\Fortify\Http\Controllers\TwoFactorAuthenticationController;
-use Laravel\Fortify\Http\Controllers\TwoFactorQrCodeController;
-use Laravel\Fortify\Http\Controllers\VerifyEmailController;
+use Naykel\Authit\Http\Controllers\UserController;
+// use Naykel\Authit\Http\Livewire\Profile;
+// use Naykel\Authit\Http\Livewire\UpdatePasswordForm;
+
+
+use Naykel\Authit\Http\Controllers\Auth\EmailVerificationNotificationController;
+use Naykel\Authit\Http\Controllers\Auth\EmailVerificationPromptController;
+use Naykel\Authit\Http\Controllers\Auth\AuthenticatedSessionController;
+use Naykel\Authit\Http\Controllers\Auth\ConfirmablePasswordController;
+use Naykel\Authit\Http\Controllers\Auth\PasswordResetLinkController;
+use Naykel\Authit\Http\Controllers\Auth\RegisterUserController;
+use Naykel\Authit\Http\Controllers\Auth\NewPasswordController;
+use Naykel\Authit\Http\Controllers\Auth\VerifyEmailController;
+
 use Spatie\Honeypot\ProtectAgainstSpam;
 
-use Naykel\Authit\Http\Controllers\UserController;
-use Naykel\Authit\Http\Livewire\Profile;
-use Naykel\Authit\Http\Livewire\UpdatePasswordForm;
+// guest users
+Route::middleware('web', 'guest')->group(function () {
 
+    Route::get('register', [RegisterUserController::class, 'create'])->middleware(ProtectAgainstSpam::class)->name('register');
+    Route::post('register', [RegisterUserController::class, 'store'])->middleware(ProtectAgainstSpam::class);
 
-Route::middleware(['web', 'auth', 'verified'])->group(function () {
+    Route::get('login', [AuthenticatedSessionController::class, 'create'])->name('login');
+    Route::post('login', [AuthenticatedSessionController::class, 'store']);
 
-    // registered user routes
-    Route::prefix('user')->name('user')->group(function () {
-        Route::get('/dashboard', [UserController::class, 'dashboard'])->name('.dashboard');
-        Route::get('/edit-account', [UserController::class, 'edit'])->name('.edit');
-        Route::get('/edit-profile', Profile::class)->name('.edit-password'); // livewire component
-        Route::get('/update-password', UpdatePasswordForm::class)->name('.update-password'); // livewire component
+    Route::get('forgot-password', [PasswordResetLinkController::class, 'create'])->name('password.request');
+    Route::post('forgot-password', [PasswordResetLinkController::class, 'store'])->name('password.email');
 
-    });
+    Route::get('reset-password/{token}', [NewPasswordController::class, 'create'])->name('password.reset');
+    Route::post('reset-password', [NewPasswordController::class, 'store'])->name('password.update');
+});
+
+// authenticated and verified users
+Route::middleware(['web', 'auth', 'verified'])->prefix('user')->name('user')->group(function () {
+
+    Route::get('/dashboard', [UserController::class, 'dashboard'])->name('.dashboard');
+    // Route::get('/edit-account', [UserController::class, 'edit'])->name('.edit');
+    // Route::get('/edit-profile', Profile::class)->name('.edit-password'); // livewire component
+    // Route::get('/update-password', UpdatePasswordForm::class)->name('.update-password'); // livewire component
+
+});
+
+// authenticated users
+Route::middleware(['web', 'auth'])->group(function () {
+
+    Route::get('verify-email', [EmailVerificationPromptController::class, '__invoke'])->name('verification.notice');
+
+    Route::get('verify-email/{id}/{hash}', [VerifyEmailController::class, '__invoke'])
+        ->middleware(['signed', 'throttle:6,1'])->name('verification.verify');
+
+    Route::post('email/verification-notification', [EmailVerificationNotificationController::class, 'store'])
+        ->middleware('throttle:6,1')->name('verification.send');
+
+    Route::get('confirm-password', [ConfirmablePasswordController::class, 'show'])->name('password.confirm');
+    Route::post('confirm-password', [ConfirmablePasswordController::class, 'store']);
+
+    Route::post('logout', [AuthenticatedSessionController::class, 'destroy'])->name('logout');
 });
 
 
-Route::group(['middleware' => config('fortify.middleware', ['web'])], function () {
-    $enableViews = config('fortify.views', true);
+// FORTIFY
 
-    // Authentication...
-    if ($enableViews) {
-        Route::get('/login', [AuthenticatedSessionController::class, 'create'])
-            ->middleware(['guest:' . config('fortify.guard')])
-            ->name('login');
-    }
+// Route::group(['middleware' => config('fortify.middleware', ['web'])], function () {
+//     $enableViews = config('fortify.views', true);
 
-    $limiter = config('fortify.limiters.login');
-    $twoFactorLimiter = config('fortify.limiters.two-factor');
-    $verificationLimiter = config('fortify.limiters.verification', '6,1');
+//     $limiter = config('fortify.limiters.login');
+//     $twoFactorLimiter = config('fortify.limiters.two-factor');
+//     $verificationLimiter = config('fortify.limiters.verification', '6,1');
 
-    Route::post('/login', [AuthenticatedSessionController::class, 'store'])
-        ->middleware(array_filter([
-            'guest:' . config('fortify.guard'),
-            $limiter ? 'throttle:' . $limiter : null,
-        ]));
+//     // Two Factor Authentication...
+//     if (Features::enabled(Features::twoFactorAuthentication())) {
+//         if ($enableViews) {
+//             Route::get('/two-factor-challenge', [TwoFactorAuthenticatedSessionController::class, 'create'])
+//                 ->middleware(['guest:' . config('fortify.guard')])
+//                 ->name('two-factor.login');
+//         }
 
-    Route::post('/logout', [AuthenticatedSessionController::class, 'destroy'])
-        ->name('logout');
+//         Route::post('/two-factor-challenge', [TwoFactorAuthenticatedSessionController::class, 'store'])
+//             ->middleware(array_filter([
+//                 'guest:' . config('fortify.guard'),
+//                 $twoFactorLimiter ? 'throttle:' . $twoFactorLimiter : null,
+//             ]));
 
-    // Password Reset...
-    if (Features::enabled(Features::resetPasswords())) {
-        if ($enableViews) {
-            Route::get('/forgot-password', [PasswordResetLinkController::class, 'create'])
-                ->middleware(['guest:' . config('fortify.guard')])
-                ->name('password.request');
+//         $twoFactorMiddleware = Features::optionEnabled(Features::twoFactorAuthentication(), 'confirmPassword')
+//             ? [config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard'), 'password.confirm']
+//             : [config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')];
 
-            Route::get('/reset-password/{token}', [NewPasswordController::class, 'create'])
-                ->middleware(['guest:' . config('fortify.guard')])
-                ->name('password.reset');
-        }
+//         Route::post('/user/two-factor-authentication', [TwoFactorAuthenticationController::class, 'store'])
+//             ->middleware($twoFactorMiddleware)
+//             ->name('two-factor.enable');
 
-        Route::post('/forgot-password', [PasswordResetLinkController::class, 'store'])
-            ->middleware(['guest:' . config('fortify.guard')])
-            ->name('password.email');
+//         Route::delete('/user/two-factor-authentication', [TwoFactorAuthenticationController::class, 'destroy'])
+//             ->middleware($twoFactorMiddleware)
+//             ->name('two-factor.disable');
 
-        Route::post('/reset-password', [NewPasswordController::class, 'store'])
-            ->middleware(['guest:' . config('fortify.guard')])
-            ->name('password.update');
-    }
+//         Route::get('/user/two-factor-qr-code', [TwoFactorQrCodeController::class, 'show'])
+//             ->middleware($twoFactorMiddleware)
+//             ->name('two-factor.qr-code');
 
+//         Route::get('/user/two-factor-recovery-codes', [RecoveryCodeController::class, 'index'])
+//             ->middleware($twoFactorMiddleware)
+//             ->name('two-factor.recovery-codes');
 
-    Route::middleware(ProtectAgainstSpam::class)->group(function () use ($enableViews) {
-        // Registration...
-        if (Features::enabled(Features::registration())) {
-            if ($enableViews) {
-                Route::get('/register', [RegisteredUserController::class, 'create'])
-                    ->middleware(['guest:' . config('fortify.guard')])
-                    ->name('register');
-            }
-
-            Route::post('/register', [RegisteredUserController::class, 'store'])
-                ->middleware(['guest:' . config('fortify.guard')]);
-        }
-    });
-
-
-
-    // Email Verification...
-    if (Features::enabled(Features::emailVerification())) {
-        if ($enableViews) {
-            Route::get('/email/verify', [EmailVerificationPromptController::class, '__invoke'])
-                ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')])
-                ->name('verification.notice');
-        }
-
-        Route::get('/email/verify/{id}/{hash}', [VerifyEmailController::class, '__invoke'])
-            ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard'), 'signed', 'throttle:' . $verificationLimiter])
-            ->name('verification.verify');
-
-        Route::post('/email/verification-notification', [EmailVerificationNotificationController::class, 'store'])
-            ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard'), 'throttle:' . $verificationLimiter])
-            ->name('verification.send');
-    }
-
-    // Profile Information...
-    if (Features::enabled(Features::updateProfileInformation())) {
-        Route::put('/user/profile-information', [ProfileInformationController::class, 'update'])
-            ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')])
-            ->name('user-profile-information.update');
-    }
-
-    // Passwords...
-    if (Features::enabled(Features::updatePasswords())) {
-        Route::put('/user/password', [PasswordController::class, 'update'])
-            ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')])
-            ->name('user-password.update');
-    }
-
-    // Password Confirmation...
-    if ($enableViews) {
-        Route::get('/user/confirm-password', [ConfirmablePasswordController::class, 'show'])
-            ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')])
-            ->name('password.confirm');
-    }
-
-    Route::get('/user/confirmed-password-status', [ConfirmedPasswordStatusController::class, 'show'])
-        ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')])
-        ->name('password.confirmation');
-
-    Route::post('/user/confirm-password', [ConfirmablePasswordController::class, 'store'])
-        ->middleware([config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')]);
-
-    // Two Factor Authentication...
-    if (Features::enabled(Features::twoFactorAuthentication())) {
-        if ($enableViews) {
-            Route::get('/two-factor-challenge', [TwoFactorAuthenticatedSessionController::class, 'create'])
-                ->middleware(['guest:' . config('fortify.guard')])
-                ->name('two-factor.login');
-        }
-
-        Route::post('/two-factor-challenge', [TwoFactorAuthenticatedSessionController::class, 'store'])
-            ->middleware(array_filter([
-                'guest:' . config('fortify.guard'),
-                $twoFactorLimiter ? 'throttle:' . $twoFactorLimiter : null,
-            ]));
-
-        $twoFactorMiddleware = Features::optionEnabled(Features::twoFactorAuthentication(), 'confirmPassword')
-            ? [config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard'), 'password.confirm']
-            : [config('fortify.auth_middleware', 'auth') . ':' . config('fortify.guard')];
-
-        Route::post('/user/two-factor-authentication', [TwoFactorAuthenticationController::class, 'store'])
-            ->middleware($twoFactorMiddleware)
-            ->name('two-factor.enable');
-
-        Route::delete('/user/two-factor-authentication', [TwoFactorAuthenticationController::class, 'destroy'])
-            ->middleware($twoFactorMiddleware)
-            ->name('two-factor.disable');
-
-        Route::get('/user/two-factor-qr-code', [TwoFactorQrCodeController::class, 'show'])
-            ->middleware($twoFactorMiddleware)
-            ->name('two-factor.qr-code');
-
-        Route::get('/user/two-factor-recovery-codes', [RecoveryCodeController::class, 'index'])
-            ->middleware($twoFactorMiddleware)
-            ->name('two-factor.recovery-codes');
-
-        Route::post('/user/two-factor-recovery-codes', [RecoveryCodeController::class, 'store'])
-            ->middleware($twoFactorMiddleware);
-    }
-});
+//         Route::post('/user/two-factor-recovery-codes', [RecoveryCodeController::class, 'store'])
+//             ->middleware($twoFactorMiddleware);
+//     }
+// });